@@ -30,16 +30,18 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
+
   http: ^1.2.1
   dio: ^5.4.3
   image_picker: ^1.1.2
   shared_preferences: ^2.2.3
-  webview_flutter: ^4.2.1
-  webview_flutter_android: ^3.9.1
-  # The following adds the Cupertino Icons font to your application.
-  # Use with the CupertinoIcons class for iOS style icons.
+
+  webview_flutter: ^4.9.0
+  webview_flutter_android: ^3.13.1
+  webview_flutter_wkwebview: ^3.13.1
+
   cupertino_icons: ^1.0.8
-  webview_flutter_wkwebview: ^3.7.1
+
 
 dev_dependencies:
   flutter_test:
